backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
anthropic>=0.34.2
python-dotenv>=1.0.1
pydantic>=2.10.0
PyJWT[cryptography]>=2.9.0
httpx>=0.27.0
supabase>=2.10.0
